// frontend/src/app/apiClient/model/company.ts
/**
 * LoopBack Application
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */






export interface Company { 

    
    companyID: string;

    
    name: string;

    
    emailForInvoices: string;

    
    emailForUpdates1?: string;

    
    emailForUpdates2?: string;

    
    emailForUpdates3?: string;

}