// frontend/src/app/apiClient/model/user.ts
/**
 * LoopBack Application
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */



export interface User { 

    
    userID: string;

    
    name: string;

    
    email: string;

    
    password: string;


    phoneNumber: string;

    
    role: string;

    
    key2FA?: string;

    
    emailVerified?: boolean;


    companyName?: string;

}